Renaming roster to flexible_roster to avoid confusion with the existing roster for the homeroom page.
