+++ app/assets/javascripts/components/flexible_roster.jsx
