+++ app/assets/stylesheets/components/flexible_roster.scss
