+++ app/assets/stylesheets/style.scss
@@ -10,7 +10,7 @@
 @import 'components/fixed_table';
 @import 'components/students_table';
 @import 'components/help_bubble';
-@import 'components/roster';
+@import 'components/flexible_roster';
 
 @import 'roster';
 @import 'about';
